User login working
Along creating accounts, logging in is now possible. The system doesn't yet keep track of if you're logged in or not, though... You can log in, it'll tell you you did so, and then nothing will have changed.

Secondly, passwords are now stored as hashes of both a static salt, a random salt, a userspecific salt, and the password itself. This should in theory make it more secure.