Logging in works
Program keeps track of if you're logged in in main.js. This is not exposed to the main program, but certain functionality (like retrieving the id you're logged in as) is exposed.

You can now only add listings for the user you're logged in as. Whom you're logged in as is also displayed.