Fix user desync
A lot of ways to fix this ended up being impractical. I'll list some of them and why they didn't work here in case its usefull for future reference:

The problem was that the function that synced the user was only being passed to defineNuxtRouteMiddleware which calls a function when you change page or something or other. This obviously means if you don't swap pages, the user desyncs untill you do which is bad.

To get around this I tried making an implementation where you'd export a function to sync them again, and call that in the backend whenever a change was made that would make cause for a desync. This ended up being impossible as this needs to be done clientside, not serverside for various reasons.

Calling it manually in the frontend whenever necessary is probably the best solution, and can still be done, but this fix makes it less of an issue if it's forgotten or omitted because the use state is automatically synced periodically no matter where you are in the app.

My (admittedly limited) testing shows that even at low time intervals (like 100ms), this is not as innefficient of resources as one might think. The user is a very small object so syncing it is no big deal. If large performance detriments are noticed, we can always remove this and simply call the SyncUser(); after any update sent to the backend that would mean it is prudent.

Finally and IMPORTANTLYLY: Note that security is not compromised by a desynced userstate, otherwise this solution would be unacceptable. It merely affects some UI elements